+++ src/main/java/com/emc/ecs/management/sdk/BucketQuotaAction.java
@@ -1,6 +1,6 @@
 package com.emc.ecs.management.sdk;
 
-import com.emc.ecs.servicebroker.EcsManagementClientException;
+import com.emc.ecs.servicebroker.exception.EcsManagementClientException;
 import com.emc.ecs.management.sdk.model.BucketQuotaDetails;
 import com.emc.ecs.management.sdk.model.BucketQuotaParam;
