+++ src/main/java/com/emc/ecs/management/sdk/BucketRetentionAction.java
@@ -1,6 +1,6 @@
 package com.emc.ecs.management.sdk;
 
-import com.emc.ecs.servicebroker.EcsManagementClientException;
+import com.emc.ecs.servicebroker.exception.EcsManagementClientException;
 import com.emc.ecs.management.sdk.model.DefaultBucketRetention;
 import com.emc.ecs.management.sdk.model.DefaultBucketRetentionUpdate;
