+++ src/main/java/com/emc/ecs/management/sdk/NamespaceQuotaAction.java
@@ -1,6 +1,6 @@
 package com.emc.ecs.management.sdk;
 
-import com.emc.ecs.servicebroker.EcsManagementClientException;
+import com.emc.ecs.servicebroker.exception.EcsManagementClientException;
 import com.emc.ecs.management.sdk.model.NamespaceQuotaDetails;
 import com.emc.ecs.management.sdk.model.NamespaceQuotaParam;
