+++ src/main/java/com/emc/ecs/servicebroker/exception/EcsManagementClientException.java
@@ -1,10 +1,11 @@
-package com.emc.ecs.servicebroker;
+package com.emc.ecs.servicebroker.exception;
 
+import org.springframework.cloud.servicebroker.exception.ServiceBrokerException;
 import org.springframework.http.HttpStatus;
 import org.springframework.web.bind.annotation.ResponseStatus;
 
 @ResponseStatus(code = HttpStatus.INTERNAL_SERVER_ERROR)
-public class EcsManagementClientException extends Exception {
+public class EcsManagementClientException extends ServiceBrokerException {
     private static final long serialVersionUID = 1L;
 
     public EcsManagementClientException(String message) {
