+++ src/test/java/com/emc/ecs/management/sdk/BucketRetentionActionTest.java
@@ -1,7 +1,7 @@
 package com.emc.ecs.management.sdk;
 
-import com.emc.ecs.servicebroker.EcsManagementClientException;
-import com.emc.ecs.servicebroker.EcsManagementResourceNotFoundException;
+import com.emc.ecs.servicebroker.exception.EcsManagementClientException;
+import com.emc.ecs.servicebroker.exception.EcsManagementResourceNotFoundException;
 import com.emc.ecs.common.EcsActionTest;
 import com.emc.ecs.management.sdk.model.DefaultBucketRetention;
 import org.junit.After;
