+++ src/test/java/com/emc/ecs/management/sdk/NFSExportActionTest.java
@@ -1,6 +1,6 @@
 package com.emc.ecs.management.sdk;
 
-import com.emc.ecs.servicebroker.EcsManagementClientException;
+import com.emc.ecs.servicebroker.exception.EcsManagementClientException;
 import com.emc.ecs.common.EcsActionTest;
 import com.emc.ecs.management.sdk.model.NFSExport;
 import org.junit.After;
