+++ src/test/java/com/emc/ecs/management/sdk/ObjectUserActionTest.java
@@ -1,6 +1,6 @@
 package com.emc.ecs.management.sdk;
 
-import com.emc.ecs.servicebroker.EcsManagementClientException;
+import com.emc.ecs.servicebroker.exception.EcsManagementClientException;
 import com.emc.ecs.common.EcsActionTest;
 import org.junit.After;
 import org.junit.Before;
